extract TrackError
