+++ extensions/blocks/podcast-player/components/playlist.js
@@ -7,37 +7,16 @@ import classnames from 'classnames';
  * WordPress dependencies
  */
 import { memo } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import TrackIcon from './track-icon';
+import TrackError from './track-error';
 import { STATE_ERROR, STATE_PLAYING } from '../constants';
 
 import { getColorClassName } from '../utils';
-
-const TrackError = memo( ( { link, title } ) => (
-	<div className="jetpack-podcast-player__track-error">
-		{ __( 'Episode unavailable. ', 'jetpack' ) }
-		{ link && (
-			<span>
-				<a href={ link } rel="noopener noreferrer nofollow" target="_blank">
-					<span className="jetpack-podcast-player--visually-hidden">
-						{ /* Intentional trailing space outside of the translated string. */ }
-						{ `${ sprintf(
-							/* translators: %s is the title of the track. This text is
-							visually hidden from the screen, but available to screen readers. */
-							__( '%s:', 'jetpack' ),
-							title
-						) } ` }
-					</span>
-					{ __( 'Open in a new tab', 'jetpack' ) }
-				</a>
-			</span>
-		) }
-	</div>
-) );
 
 const Track = memo(
 	( {
